test: Tests taking into account rental overlapping
