+++ controllers/SpaceControllerService.js
@@ -447,7 +447,7 @@ module.exports.postSpaceRental = async function postSpaceRental (req, res, next)
         res.status(400).send('Cannot rent space twice. Please update or delete your previous rental of this space');
         return;
       }
-      const errors = utils.checkRentalValidity(rentalToBeCreated, spaceToAddRental);
+      const errors = utils.rental.checkRentalValidity(rentalToBeCreated, spaceToAddRental);
       if (errors.length > 0) {
         res.status(400).send(`Bad Request: ${errors[0]}`);
         return;
@@ -482,7 +482,6 @@ module.exports.postSpaceRental = async function postSpaceRental (req, res, next)
       if (err instanceof jwt.JsonWebTokenError) {
         res.status(401).send(`Unauthorized: ${err.message}`);
       } else {
-        console.error(err);
         res.status(500).send('Internal Server Error');
       }
     }
